sa: truncate all timestamps to seconds
As described in #7075, go-sql-driver/mysql v1.5.0 truncates timestamps to
microseconds, while v1.6.0 and above does not. That means upon upgrading to
v1.6.0, timestamps are written to the database with a resolution of nanoseconds,
and SELECT statements also use a resolution of nanoseconds. We believe this is
the cause of performance problems we observed when upgrading to v1.6.0 and
above.

To fix that, apply rounding in the application code. Rather than just rounding
to microseconds, round to seconds since that is the resolution we care about.
Using seconds rather than microseconds may also allow some of our indexes to
grow more slowly over time.